T05
- FourBoards로 큐브 코드 만드는 것 분리하여 더 모듈화 진행
